+++ src/main/java/com/example/dorywcza/repository/ServiceOfferRepository.java
@@ -3,5 +3,5 @@ package com.example.dorywcza.repository;
 import com.example.dorywcza.model.service_offer.ServiceOffer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceOfferRepository extends JpaRepository<ServiceOffer, Integer> {
+public interface ServiceOfferRepository extends JpaRepository<ServiceOffer, Long> {
 }
